requirements.txt
pytorch-lightning
learn2learn >= 0.1.6
comet-ml
resemblyzer
g2p-en
inflect
librosa
matplotlib
numba
numpy
pypinyin
pyworld
PyYAML
scikit-learn
scipy
soundfile
tensorboard
tgt
torch
tqdm
unidecode
torchaudio

# Evaluation
seaborn
-e git+https://github.com/jasminsternkopf/mel_cepstral_distance.git@main#egg=mcd
-e git+https://github.com/aliutkus/speechmetrics#egg=speechmetrics[gpu]
